Do track Eclipse pydev files after all
Previously I thought these should not have been versioned, but I changed my
mind. It is unnecessary to set them up every time for a pydev project in Eclipse
and it's ok to version them, as long as we keep track that everything commited
is using variables and is not specific to one developer's environment.